// frontend/src/domain/enums/enum.ts
enum UserRole {
  ADMIN = "ADMIN",
  ACCOUNTANT = "ACCOUNTANT",
  LOGISTIC_WORKER = "LOGISTIC_WORKER",
  SECURITY = "SECURITY",
}

enum DocumentType {
  POD = "POD",
  LR_COPY = "LR_COPY",
  INVOICE_PDF = "INVOICE_PDF",
  VEHICLE_PHOTO = "VEHICLE_PHOTO",
  PAYMENT_PROOF = "PAYMENT_PROOF",
  BENEFICIARY_DOC = "BENEFICIARY_DOC",
}

export { DocumentType, UserRole };